Move unnecessary functions from extension.ts
